// src/app/page/master/pemilik-profile/pemilik-profile.interface.ts
import { EventEmitter } from '@angular/core';

export interface PemilikProfile {
    kode?: any;
    namaPemilikProfile?: string;
    reportDisplay?: string;
    kodeExternal?: string;
    namaExternal?: string;
    kdKelompokPemilik?: boolean;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    kelompokPemilik?: string;
    id?: any;
}